# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jef Bot API",
    "version": "1.0.0",
    "summary": "AI chat moderation for Discord servers, game chat and any app.",
    "description": "The /v1 API of Jef Bot. Authenticate with `Authorization: Bearer <key>`: account keys (jef_acct_…) can do everything on the account; app keys (jef_app_…) can only moderate and read one app (plus warn/pardon its users). Errors are `{ \"error\": { \"code\", \"message\" } }`. Timestamps are ISO 8601 UTC. Guides: /docs/agents.html, /docs/quickstart-game.html. Full text for LLMs: /llms-full.txt.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This Jef Bot server"
    }
  ],
...
  "security": [
    {
      "accountKey": []
    },
    {
      "appKey": []
    },
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/v1/billing": {
      "get": {
        "operationId": "getBilling",
        "tags": [
          "Billing"
        ],
        "summary": "Billing status",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Billing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Billing"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          }
        }
      }
    },
    "/v1/billing/checkout": {
      "post": {
        "operationId": "createCheckout",
        "tags": [
          "Billing"
        ],
        "summary": "Stripe Checkout URL for a person to subscribe (or the portal if already subscribed)",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              },
              "example": {
                "plan": "starter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "502": {
            "$ref": "#/components/responses/E502"
          },
          "503": {
            "$ref": "#/components/responses/E503"
          }
        }
      }
    },
    "/v1/billing/portal": {
      "post": {
        "operationId": "createPortalSession",
        "tags": [
          "Billing"
        ],
        "summary": "Stripe customer portal URL (change plan, card, invoices, cancel)",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "502": {
            "$ref": "#/components/responses/E502"
          },
          "503": {
            "$ref": "#/components/responses/E503"
          }
        }
      }
    }
  },
...
  "components": {
    "securitySchemes": {
      "accountKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jef_acct_…",
        "description": "Account key: full access to the account."
      },
      "appKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jef_app_…",
        "description": "App key: moderation and reads for one app."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jef_session",
        "description": "Dashboard session after Google sign-in. Changes must be same-origin JSON."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable code, e.g. invalid_request, unauthorized, invalid_api_key, forbidden, plan_limit, limit, not_found, guild_taken, rate_limited, invalid_json, too_large, bad_origin, json_required, no_customer, billing_unavailable, billing_error, internal."
              },
              "message": {
                "type": "string",
                "description": "Human-readable explanation."
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Billing": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "starter",
              "pro",
              "internal"
            ]
          },
          "subscriptionStatus": {
            "type": [
              "string",
              "null"
            ]
          },
          "currentPeriodEnd": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "hasCustomer": {
            "type": "boolean"
          }
        },
        "required": [
          "enabled",
          "plan",
          "subscriptionStatus",
          "currentPeriodEnd",
          "hasCustomer"
        ]
      },
      "CheckoutRequest": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "starter",
              "pro"
            ]
          }
        },
        "required": [
          "plan"
        ]
      },
      "CheckoutResult": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "A Stripe page for a person to open."
          },
          "kind": {
            "type": "string",
            "enum": [
              "checkout",
              "portal"
            ],
            "description": "portal when the account already has a subscription."
          }
        },
        "required": [
          "url",
          "kind"
        ]
      },
      "PortalResult": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "url"
        ]
      },
...
    },
    "responses": {
      "E400": {
        "description": "invalid_request, invalid_json (or no_customer for the billing portal)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E401": {
        "description": "unauthorized (no credentials) or invalid_api_key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E403": {
        "description": "forbidden (app key where an account is needed), plan_limit (over a plan limit, an app with inPlan false, or an account key past the plan's key count), limit, or bad_origin",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "E502": {
        "description": "billing_error (Stripe failed)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E503": {
        "description": "billing_unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
  }
}